requirements.txt
flask>=3.1.1
bson>=0.5.10
pymongo>=4.0
flask-pymongo>=3.0.1
pydantic>=2.10.6
pyyaml>=6.0.2
markupsafe>=3.0.2
werkzeug>=3.1.3
networkx>=3.4.2